feat: large memory support, scheduler hardening, and untyped provisioning
Kernel:
- Dynamic direct physical map sized to actual RAM (up to 512 GB)
- Two-phase frame allocator init: bitmap in Phase 1 (identity-mapped),
  per-frame tracking arrays in Phase 2 (after direct map established)
- Untyped watermark widened from u32 to u64 for >4GB regions
- MAX_SIZE_BITS raised from 28 to 38 for large untyped creation
- Kernel stacks enlarged from 4 KiB to 16 KiB (init + TCB configure)
- CLI in usermode_trampoline before CR3 switch to prevent IRQ on
  half-built ring-3 frame

Scheduler:
- Ready queue double-insertion guard via is_ready_queued_unlocked()
- schedule_unlocked loops to skip stale/running TCBs
- IRQ restore moved after context switch across all call sites
- Fastpath same-CPU check tightened: validates last_cpu and pending_cpu

IPC:
- recv() and recv_timeout() clear stale reply_tcb to prevent PIP leak

Userland:
- mmsrv: dynamic FREE_SLOTS/FRAME_POOL arrays sized by system RAM,
  MmRegion/ShmObject fields widened u16->u32, lazy procmgr EP lookup
  for sub-untyped provisioning when local sources exhausted
- procmgr: PM_REQUEST_UNTYPED handler for mmsrv sub-untyped provisioning,
  UT_MIRROR_COUNT 8->16, MAX_UT_SOURCES 12->18
- init: UT_MIRROR_COUNT 8->16
- blkdrv: drain stale virtio completions, increase timeout, ISR ack on
  timeout

+++ userland/core/mmsrv/src/client.rs
@@ -171,7 +171,7 @@ pub(crate) fn clear_cow_bit(region: *mut MmRegion, page_idx: usize) {
 #[derive(Clone, Copy)]
 struct BitmapFreeEntry {
     ptr: *mut u64,
-    alloc_pages: u16,
+    alloc_pages: u32,
 }
 
 const BITMAP_POOL_MAX: usize = 32;
@@ -188,7 +188,7 @@ static mut BITMAP_POOL_COUNT: usize = 0;
 ///
 /// Checks the free-list pool first for a suitably-sized entry;
 /// falls through to self_mmap if none available.
-pub(crate) fn alloc_cow_bitmap(page_count: usize) -> (*mut u64, u16) {
+pub(crate) fn alloc_cow_bitmap(page_count: usize) -> (*mut u64, u32) {
     let word_count = (page_count + 63) / 64;
     let byte_count = word_count * 8;
     let alloc_pages = (byte_count + 4095) / 4096;
@@ -199,7 +199,7 @@ pub(crate) fn alloc_cow_bitmap(page_count: usize) -> (*mut u64, u16) {
         let count = *(&raw const BITMAP_POOL_COUNT);
         let pool = &raw mut BITMAP_POOL;
         for i in 0..count {
-            if (*pool)[i].alloc_pages >= alloc_pages as u16 {
+            if (*pool)[i].alloc_pages >= alloc_pages as u32 {
                 let entry = (*pool)[i];
                 // Swap-remove: move last entry into this slot
                 let last = count - 1;
@@ -210,7 +210,7 @@ pub(crate) fn alloc_cow_bitmap(page_count: usize) -> (*mut u64, u16) {
                 // Zero the reused memory
                 // SAFETY: entry.ptr is a valid allocation of entry.alloc_pages pages.
                 core::ptr::write_bytes(entry.ptr as *mut u8, 0, (entry.alloc_pages as usize) * 4096);
-                return (entry.ptr, word_count as u16);
+                return (entry.ptr, word_count as u32);
             }
         }
     }
@@ -221,7 +221,7 @@ pub(crate) fn alloc_cow_bitmap(page_count: usize) -> (*mut u64, u16) {
     if ptr.is_null() {
         return (core::ptr::null_mut(), 0);
     }
-    (ptr as *mut u64, word_count as u16)
+    (ptr as *mut u64, word_count as u32)
 }
 
 /// Free a COW bitmap from a region, returning the pages to the free-list pool.
@@ -249,7 +249,7 @@ pub(crate) fn free_cow_bitmap(region: *mut MmRegion) {
         if count < BITMAP_POOL_MAX {
             (*(&raw mut BITMAP_POOL))[count] = BitmapFreeEntry {
                 ptr,
-                alloc_pages: alloc_pages as u16,
+                alloc_pages: alloc_pages as u32,
             };
             *(&raw mut BITMAP_POOL_COUNT) = count + 1;
         }
